refactor: push common vector-search parameters to a base class
